Make initiating git command available inside hook
Exports the git command that initiated the execution of the hook
available to the hook's code as `process.env.GIT_COMMAND`.

This can be accessed in regular git hooks through `ps -ocommand= -p $PPID`,
but is not available when running hook through an npm script.

It has quite a few use cases. For instance, when guarding against force
pushes in a `pre-push` hook, we can check if `process.env.GIT_COMMAND`
includes `--force ` or `-f`.