Update index.ts
The way we all use husky is by "npm install".

It is very annoying when for a reason, the hooks does not install. Then we have to go to every developers to ask them to check their install. Which is very not convenient.

This proposal, is to remove one fail, which is, in my opinion, not so difficult to handle.
Is ".git" exists, but ".git/hooks" does not, instead of skipping installation, we create the folder. Not a big deal, but sooooo annoying the current way.