don't default to run-node, use 'node' if it works
If `node` works, we will not bother with `run-node` on non-windows
platforms, since `run-node` requires `bash` and operating systems like
OpenBSD and FreeBSD, as well as minimal docker images like Alpine,
don't have bash in their base installations.

I have clarified that run-node is a fallback, not default, and updated
the snap.